Changed all fetches to async

+++ src/components/PublicDeck.tsx
@@ -75,11 +75,26 @@ function PublicDeck (props: Deck_t) {
         handleClose(event);
     };
 
-    const handleDeleteDeck = () => {
+    const handleDeleteDeck = async () => {
         handleCloseAlert();
 
+        const deckRequest = await fetch(`${requestPath}/decks/?id=${props.id}`, {
+            method: 'DELETE',
+            headers: {
+                'Authorization': 'Token ' + localStorage.getItem("token"),
+            }     
+        });
+
+        if(!deckRequest.ok) {
+            return;
+        }
+       
+        console.log("UPDATE METHOD");
+
+        props.updateMethod();
+
         //TODO
-        fetch(`${requestPath}/decks/?id=${props.id}`, {
+        /*fetch(`${requestPath}/decks/?id=${props.id}`, {
             method: 'DELETE',
             headers: {
                 'Authorization': 'Token ' + localStorage.getItem("token"),
@@ -92,7 +107,7 @@ function PublicDeck (props: Deck_t) {
 
             console.log("Delete"); // TODO
             props.updateMethod();
-        });
+        });*/
     };
 
     const button_id:string = `button-${props.id}`;
